tmp: temporary changes to update simulation code
* added TODOs to mark unclear things or things to change/improve

* moved `NucleusFluorophoreDistribution` to separate module

* updated `__init__.py`
